+++ resources/measures/ResidentialGeometryCreateSingleFamilyDetached/measure.rb
@@ -707,6 +707,7 @@ class CreateResidentialSingleFamilyDetachedGeometry < OpenStudio::Measure::Model
       spaces = model.getSpaces
       spaces.each do |space|
         next unless Geometry.get_space_floor_z(space) + UnitConversions.convert(space.zOrigin, 'm', 'ft') < 0
+
         surfaces = space.surfaces
         surfaces.each do |surface|
           next if surface.surfaceType.downcase != 'wall'
@@ -741,6 +742,7 @@ class CreateResidentialSingleFamilyDetachedGeometry < OpenStudio::Measure::Model
       end
       space_with_roof_over_garage.surfaces.each do |surface|
         next unless (surface.surfaceType.downcase == 'roofceiling') && (surface.outsideBoundaryCondition.downcase == 'outdoors')
+
         n_points = []
         s_points = []
         surface.vertices.each do |vertex|
@@ -880,6 +882,7 @@ class CreateResidentialSingleFamilyDetachedGeometry < OpenStudio::Measure::Model
 
         garage_attic_space.surfaces.each do |surface|
           next unless (num_floors > 1) || (attic_type == 'finished attic')
+
           m = Geometry.initialize_transformation_matrix(OpenStudio::Matrix.new(4, 4, 0))
           m[2, 3] = -attic_space.zOrigin
           transformation = OpenStudio::Transformation.new(m)