+++ resources/measures/ResidentialHotWaterHeaterTank/measure.rb
@@ -34,7 +34,7 @@ class ResidentialHotWaterHeaterTank < OpenStudio::Measure::ModelMeasure
 
     args = ruleset::OSArgumentVector.new
 
-    # make a string argument for furnace fuel type
+    # make a string argument for water heater fuel type
     fuel_display_names = OpenStudio::StringVector.new
     fuel_display_names << Constants.FuelTypeGas
     fuel_display_names << Constants.FuelTypeOil
@@ -54,13 +54,37 @@ class ResidentialHotWaterHeaterTank < OpenStudio::Measure::ModelMeasure
     tank_volume.setDefaultValue(Constants.Auto)
     args << tank_volume
 
+    # make an argument for setpoint type (constant or variable)
+    setpoint_type_args = OpenStudio::StringVector.new
+    setpoint_type_args << Constants.WaterHeaterSetpointTypeConstant
+    setpoint_type_args << Constants.WaterHeaterSetpointTypeScheduled
+    setpoint_type = OpenStudio::Measure::OSArgument::makeChoiceArgument('setpoint_type', setpoint_type_args, true)
+    setpoint_type.setDisplayName('Setpoint type')
+    setpoint_type.setDescription("The water heater setpoint type. The '#{Constants.WaterHeaterSetpointTypeConstant}' option will use a constant value for the whole year, while '#{Constants.WaterHeaterSetpointTypeScheduled}' will use 8760 values in a schedule file.")
+    setpoint_type.setDefaultValue(Constants.WaterHeaterSetpointTypeConstant)
+    args << setpoint_type
+
     # make an argument for hot water setpoint temperature
     setpoint_temp = osargument::makeDoubleArgument('setpoint_temp', true)
     setpoint_temp.setDisplayName('Setpoint')
-    setpoint_temp.setDescription('Water heater setpoint temperature.')
+    setpoint_temp.setDescription('Water heater setpoint temperature. This value will be ignored if the setpoint type is Scheduled.')
     setpoint_temp.setUnits('F')
     setpoint_temp.setDefaultValue(125)
     args << setpoint_temp
+
+    # make an argument for the directory that contains the hourly schedules
+    sp_dir = OpenStudio::Measure::OSArgument.makeStringArgument('schedules_directory', true)
+    sp_dir.setDisplayName('Setpoint and Operating Mode Schedule Directory')
+    sp_dir.setDescription('Absolute (or relative) directory to schedule files. This argument will be ignored if a constant setpoint type is used instead.')
+    sp_dir.setDefaultValue('./resources')
+    args << sp_dir
+
+    # make an argument for the 8760 hourly setpoint schedule
+    sp_sch = OpenStudio::Measure::OSArgument.makeStringArgument('setpoint_schedule', true)
+    sp_sch.setDisplayName('Setpoint Schedule File Name')
+    sp_sch.setDescription('Name of the hourly setpoint schedule. Setpoint should be defined (in F) for every hour. The operating mode schedule must also be located in the same location.')
+    sp_sch.setDefaultValue('hourly_setpoint_schedule.csv')
+    args << sp_sch
 
     # make a choice argument for location
     location_args = OpenStudio::StringVector.new
@@ -68,7 +92,7 @@ class ResidentialHotWaterHeaterTank < OpenStudio::Measure::ModelMeasure
     Geometry.get_model_locations(model).each do |loc|
       location_args << loc
     end
-    location = OpenStudio::Measure::OSArgument::makeChoiceArgument('location', location_args, true, true)
+    location = OpenStudio::Measure::OSArgument::makeChoiceArgument('location', location_args, true)
     location.setDisplayName('Location')
     location.setDescription("The space type for the location. '#{Constants.Auto}' will automatically choose a space type based on the space types found in the model.")
     location.setDefaultValue(Constants.Auto)
@@ -120,6 +144,16 @@ class ResidentialHotWaterHeaterTank < OpenStudio::Measure::ModelMeasure
     has_water_heater_flue.setDefaultValue(false)
     args << has_water_heater_flue
 
+    # make a string argument for mixed or stratified tank
+    tank_model_type_display_name = OpenStudio::StringVector.new
+    tank_model_type_display_name << Constants.WaterHeaterTypeTankModelTypeMixed
+    tank_model_type_display_name << Constants.WaterHeaterTypeTankModelTypeStratified
+    tank_model_type = OpenStudio::Measure::OSArgument::makeChoiceArgument('tank_model_type', tank_model_type_display_name, true)
+    tank_model_type.setDisplayName('Tank Type')
+    tank_model_type.setDescription("Type of tank model to use. The '#{Constants.WaterHeaterTypeTankModelTypeStratified}' tank generally provide more accurate results, but may significantly increase run time.")
+    tank_model_type.setDefaultValue(Constants.WaterHeaterTypeTankModelTypeMixed)
+    args << tank_model_type
+
     return args
   end # end the arguments method
 
@@ -134,10 +168,14 @@ class ResidentialHotWaterHeaterTank < OpenStudio::Measure::ModelMeasure
     energy_factor = runner.getStringArgumentValue('energy_factor', user_arguments)
     recovery_efficiency = runner.getDoubleArgumentValue('recovery_efficiency', user_arguments)
     location = runner.getStringArgumentValue('location', user_arguments)
+    setpoint_type = runner.getStringArgumentValue('setpoint_type', user_arguments)
     setpoint_temp = runner.getDoubleArgumentValue('setpoint_temp', user_arguments).to_f
+    schedule_directory = runner.getStringArgumentValue('schedules_directory', user_arguments)
+    setpoint_schedule = runner.getStringArgumentValue('setpoint_schedule', user_arguments)
     oncycle_power = runner.getDoubleArgumentValue('oncyc_power', user_arguments)
     offcycle_power = runner.getDoubleArgumentValue('offcyc_power', user_arguments)
     model.getBuilding.additionalProperties.setFeature('has_water_heater_flue', runner.getBoolArgumentValue('has_water_heater_flue', user_arguments))
+    tank_model_type = runner.getStringArgumentValue('tank_model_type', user_arguments)
 
     # Validate inputs
     if not runner.validateUserArguments(arguments(model), user_arguments)
@@ -165,6 +203,39 @@ class ResidentialHotWaterHeaterTank < OpenStudio::Measure::ModelMeasure
     end
 
     location_hierarchy = Waterheater.get_location_hierarchy(ba_cz_name)
+
+    if setpoint_type == Constants.WaterHeaterSetpointTypeScheduled
+      unless (Pathname.new schedule_directory).absolute?
+        schedule_directory = File.expand_path(File.join(File.dirname(__FILE__), schedule_directory))
+      end
+      setpoint_schedule_file = File.join(schedule_directory, setpoint_schedule)
+
+      # Error checking: make sure the schedule file is an appropriate length and values are valid
+      # setpoint_hourly = HourlySchedule.new(model, runner, sch_name, dr_schedule_file, 0, false, [])
+      # setpoint_hourly_array = setpoint_hourly.schedule_array.map { |x| x.to_i }
+      # setpoint_hourly.schedule.remove
+
+      # year_description = model.getYearDescription
+      # assumed_year = year_description.assumedYear
+      # yr_hrs = Constants.NumHoursInYear(year_description.isLeapYear)
+      # run_period = model.getRunPeriod
+      # run_period_start = Time.new(assumed_year, run_period.getBeginMonth, run_period.getBeginDayOfMonth)
+      # run_period_end = Time.new(assumed_year, run_period.getEndMonth, run_period.getEndDayOfMonth, 24)
+      # sim_hours = (run_period_end - run_period_start) / 3600
+
+      # year_description = model.getYearDescription
+      # n_days = Constants.NumDaysInYear(year_description.isLeapYear)
+      # if (sp_hourly_array.length != sim_hours) and (sp_hourly_array.length != n_days * 24)
+      #  runner.registerError("Hourly water heater setpoint schedule length must equal to a full year")
+      #  return false
+      # end
+      # return true
+
+      if not File.exist?(setpoint_schedule_file)
+        runner.registerError("'#{setpoint_schedule_file}' does not exist.")
+        return false
+      end
+    end
 
     Waterheater.remove(model, runner)
 
@@ -194,8 +265,10 @@ class ResidentialHotWaterHeaterTank < OpenStudio::Measure::ModelMeasure
 
       success = Waterheater.apply_tank(model, unit, runner, loop, space, fuel_type,
                                        capacity, tank_volume, energy_factor,
-                                       recovery_efficiency, setpoint_temp,
-                                       oncycle_power, offcycle_power, 1.0)
+                                       recovery_efficiency, setpoint_type,
+                                       setpoint_temp, setpoint_schedule_file,
+                                       oncycle_power, offcycle_power, 1.0,
+                                       tank_model_type)
       return false if not success
     end